Made changes to allow compilation with VC++.

+++ src/game/Level.cpp
@@ -118,7 +118,7 @@ namespace jvgs
 
         int Level::getNumberOfEntities() const
         {
-            return entities.size();
+            return (int) entities.size();
         }
 
         Entity *Level::getEntity(int index) const
